updated(code): transformJSX + @babel/standaloneのコード
