+++ package.json
@@ -57,6 +57,8 @@
     }
   },
   "dependencies": {
+    "@babel/core": "^7.1.2",
+    "@babel/plugin-transform-react-jsx": "^7.0.0",
     "@babel/standalone": "^7.1.0",
     "body-parser": "^1.18.3",
     "cookie-parser": "^1.4.3",
